+++ database/init_db.development/post-init-db.sql
@@ -16,6 +16,7 @@ BEGIN
         ALTER TABLE development.shdi SET SCHEMA public;
         ALTER TABLE development.hexcells SET SCHEMA public;
         ALTER TABLE development.admin_world_water SET SCHEMA public;
+        ALTER TABLE development.mapaction_cesdrr_countries SET SCHEMA public;
     ELSE
         ALTER TABLE test.regions SET SCHEMA public;
         ALTER TABLE test.shdi SET SCHEMA public;
@@ -32,6 +33,8 @@ DROP TABLE IF EXISTS development.shdi;
 DROP TABLE IF EXISTS development.hexcells;
 
 DROP TABLE IF EXISTS development.admin_world_water;
+
+DROP TABLE IF EXISTS development.mapaction_cesdrr_countries;
 
 DROP TABLE IF EXISTS test.regions;