+++ database/scripts/development-database/remove_dev_schema.sql
@@ -20,9 +20,7 @@ DROP TABLE IF EXISTS test.hexcells;
 
 DROP TABLE IF EXISTS development.admin_world_water;
 
-DROP TABLE IF EXISTS development.mapaction_cesdrr_countires;
-
-DROP TABLE IF EXISTS test.mapaction_cesdrr_countires;
+DROP TABLE IF EXISTS development.mapaction_cesdrr_countries;
 
 DROP SCHEMA IF EXISTS development;