+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 - Improve Currentness indicator ([#274])
 - Add new report `MultilevelCurrentness` ([#403])
-- Add MapAction priority countries to database ([#427])
+- Add MapAction priority countries to database ([#427] [#428])
 
 ### Other Changes
 
@@ -19,6 +19,7 @@
 [#416]: https://github.com/GIScience/ohsome-quality-analyst/pull/416
 [#426]: https://github.com/GIScience/ohsome-quality-analyst/pull/426
 [#427]: https://github.com/GIScience/ohsome-quality-analyst/pull/427
+[#428]: https://github.com/GIScience/ohsome-quality-analyst/pull/428
 
 
 ## 0.11.0